requirements.txt
streamlit>=1.34.0
handcalc>=1.6.5
pandas>=2.2.2
forallpeople>=2.6.7
